Add initial PulseAudio support.
The user can now optionally expose themselves to PulseAudio exploits and
ultrasonic government mindcontrol beams if they choose to do so.

Note: Querying the PulseAudio properties from the X11 root window is not
supported yet.  It's relatively easy to do, but requires that I write
cgo, or pull in someone else's xlib/xcb bindings that have the cgo bits.

Fixes #5.