Deal with libgconf in the worst possible way. #32
Tor Browser's Firefox is compiled with GNOME integration, and Firefox's
idea of a good way to handle "libgconf exists, no D-Bus" is to pop up a
dialog box.

Work around this by abusing the `tbb_stub.so` shared object that gets
preloaded in, to hijack the `dlopen()` call with one that can examine
which shared objects we're interested in.