Make this play nice with grsec.
Under the assumption that TBB 7.0 will be ESR52 based, on all bundles
prior to that, the launcher will disable PaX MPROTECT for the firefox
binary if it detects a grsec kernel by setting the xattr override.

Additionally the `hardened` series is totally incompatible with grsec
kernels due to asan, so an error is displayed instead of the launch
going through.

Fixes #28.