Implement the rest of the circuit diplay code.
This adds support for the bare minimum required to get the circuit
display to function correctly with 6.5a4.  Only circuits and streams
that originate from the sandboxed Tor Browser instance will be visible
to firefox, so this should be safe to use with a system tor instance,
relatively speaking.

Note: If the Tor Browser circuit monitor changes in the way it
functions, this codw will likely need revision.

Fixes #3.