Attempt to bring sanity to the seccomp situation.
 * Use a separate seccomp whitelist crafted for tor.
 * Revert most of the changes to the profile I stole from subgraph.

The tor profile is based off tor's internal sandbox, with modifications
to add the system calls that are executed before Tor enable's it's
sandbox (thus misssing from the profile), and to fit the limitations of
the current profile parser.

Tor's built in sandbox is still preferable, but it doesn't appear that
Tor Browser's tor is actually built with seccomp support.