Add a passthrough SOCKS listener.
When sandboxed-tor-browser is managing the tor process, launch a pass
through listener on `127.0.0.1:9150` to allow other applications to use
the sandboxed tor's SOCKSPort.

This feature is disabled if a system tor is in use under the assumption
that, the can edit the torrc to do whatever they want anyway, and if
that port is already taken, it will fail gracefully (ie: log a warning).