Step 1 for fixing the library mess. #38
At launch time, examine all the binaries that will be executed in the
sandbox and use `/etc/ld.so.cache` to figure out which libraries should
be in the container.

Only works on amd64 for now, becaues the i386 library situation is
slightly more complex.  Tested on Arch, Debian stable, Fedora 25.