Re-add socket call to the 386 whitelist.
It helps to test code involving ancient bullshit that needs to die, on
systems that actually exercise said ancient bullshit.  In this case,
Debian stable x86 conveniently provides a kernel that actually
exercises `socketcall`.

libseccomp2 is supposed to "automagically do the right thing" when
generating 386 rules since I'm not adding exact, but both tor and
firefox showed problems with this.