Bug 21093: Unbreak the Debian build, by kludging gosecco.
gosecco has code to load seccomp filters with the Linux 3.17 system
call, but certain copies of the debian headers pre-date the introduction
of this.

Disabling all the calls is harmless since all gosecco is used for is
parsing and copiling rules, and the disabled routines are for using the
library to install the filter (which we offload to bubblewrap).

Note: Why yes, this is a repeat of 3b568bda74.