Enforce the patch size against that listed in the update XML metadata.
This is more defense in depth than anything else since the patch's
signature is checked, and the update XML file is what I consider
semi-trusted (fetched from a source that's cryptographically
authenticated either via HPKP or a `.onion`, but not signed on it's
own).

It is however the sensible thing to do since SHA512 is susceptible to
length-extension attacks.