Bug 21184: Do a better job of killing/cleaning up bwrap children.
Use the bwrap fork that's acting as init to kill, wait on the main bwrap
process returned from os/exec when waiting/polling process status.

This can be further improved, but would require bubblewrap patches to
make substantially better, and at least the cleanup/refactor lays the
groundwork for everything.