Call setsid() before we exec bwrap.
This is part of a defense in depth measure to `TOCSTI` considered
harmful (CVE-2017-5226).  Released versions of `sandboxed-tor-browser`
was NEVER vulnerable to the issue at all because the seccomp whitelist
does not allow `TIOCSTI`.

bubblewrap's behavior here changes a bit depending on the version and
backports.  In theory if everyone shipped the latest and greatest
(master), `--new-session` is available as an option, but I can just
do this myself.