Bug 20773: Stop mounting /proc in the Tor Browser container.
Wasn't so bad to figure out, though upstream still should fall back
gracefully, since the only reason why it crashes horribly is because
pthread_attr_getstack() returns a stack size of 0 for the default thread
when /proc is missing.

In our case we can safely fall back to querying RLIMIT_STACK, because
the sandbox containerization explicity sets it to something large, but
relatively sane.