Bug 22932: Add experimental Amensiac Profile Directory support.
This is currently an advanced config option as it is still experimental
and needs more testing.  When enabled it will copy the Firefox profile
directory into a tmpfs mount when the container is created, that is
discarded when the container is torn down.

The following directories are read-only bind mounted as usual:
 * profile.default/preferences
 * profile.default/extensions