Unset the addon autoupdater URL prefs.
Updating addons is disabled by default and will not work, even though it
may appear to (#23265).  Instead of checking and then failing to apply
the update, by removing the URLs from the prefs, the check will fail
immediately.

This applies to everything but HTTPSE, because HTTPSE provides it's own
updateURL in the install.rdf file.